requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# CARLA (installer séparément)
# carla==0.9.13

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualisation
matplotlib>=3.7.0
tensorboard>=2.13.0

# Utilitaires
numpy>=1.24.0
tqdm>=4.65.0

# Optionnel mais recommandé
scikit-learn>=1.3.0
scipy>=1.11.0